various fixes and refactoring
