+++ PandaTechEShop/Utilities/Dialog/DialogService.cs
@@ -1,0 +1,7 @@
+namespace PandaTechEShop.Utilities.Dialog
+{
+    public class DialogService
+    {
+        
+    }
+}
