+++ PandaTechEShop/Utilities/Dialog/IDialogService.cs
@@ -1,0 +1,7 @@
+namespace PandaTechEShop.Utilities.Dialog
+{
+    public interface IDialogService
+    {
+        
+    }
+}
