+++ PandaTechEShop/Utilities/SecureStorage/ISecureStorageService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+
+namespace PandaTechEShop.Services.SecureStorage
+{
+    public interface ISecureStorageService
+    {
+        Task<bool> SetAsync(string key, string value);
+        Task<string> GetAsync(string key);
+        Task<bool> Remove(string key);
+    }
+}
